backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.13
httpx==0.28.1
anthropic==0.43.0
apscheduler==3.10.4
slowapi==0.1.9
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1